+++ py/desitarget/test/test_geomask.py
@@ -3,7 +3,6 @@
 """Test desitarget.geomask.
 """
 import unittest
-from pkg_resources import resource_filename
 import numpy as np
 import os
 
@@ -57,14 +56,3 @@ class TestGEOMASK(unittest.TestCase):
         self.assertTrue(np.all(a[iia] == b[iib]))
         self.assertTrue(len(iia), len(a))
 
-
-if __name__ == '__main__':
-    unittest.main()
-
-
-def test_suite():
-    """Allows testing of only this module with the command:
-
-        python setup.py test -m desitarget.test.test_geomask
-    """
-    return unittest.defaultTestLoader.loadTestsFromName(__name__)