+++ py/desitarget/test/test_mtl.py
@@ -406,14 +406,3 @@ class TestMTL(unittest.TestCase):
                                 "ZWARN bit value mismatch for {}".format(
                                     bitname))
 
-
-if __name__ == '__main__':
-    unittest.main()
-
-
-def test_suite():
-    """Allows testing of only this module with the command:
-
-        python setup.py test -m desitarget.test.test_mtl
-    """
-    return unittest.defaultTestLoader.loadTestsFromName(__name__)