+++ py/desitarget/test/test_mtl_multiple.py
@@ -128,14 +128,3 @@ class TestMTL(unittest.TestCase):
             toohigh |= more_zwarn > done
             self.assertFalse(toomany & toohigh, msg.format(bitname))
 
-
-if __name__ == '__main__':
-    unittest.main()
-
-
-def test_suite():
-    """Allows testing of only this module with the command:
-
-        python setup.py test -m desitarget.test.test_mtl
-    """
-    return unittest.defaultTestLoader.loadTestsFromName(__name__)